Move Meeting Sidebar component to SidePanel
