+++ modules/meeting/app/components/meetings/side_panel/attachments_component.rb
@@ -27,55 +27,15 @@
 #++
 
 module Meetings
-  class Sidebar::ParticipantsComponent < ApplicationComponent
+  class SidePanel::AttachmentsComponent < ApplicationComponent
     include ApplicationHelper
     include OpTurbo::Streamable
     include OpPrimer::ComponentHelpers
-
-    MAX_SHOWN_PARTICIPANTS = 5
-
-    def wrapper_data_attributes
-      {
-        controller: "expandable-list",
-        "application-target": "dynamic"
-      }
-    end
 
     def initialize(meeting:)
       super
 
       @meeting = meeting
-    end
-
-    def elements
-      @elements ||= @meeting.invited_or_attended_participants.sort
-    end
-
-    def count
-      @count ||= elements.count
-    end
-
-    def render_participant(participant)
-      flex_layout(align_items: :center) do |flex|
-        flex.with_column(classes: "ellipsis") do
-          render(Users::AvatarComponent.new(user: participant.user,
-                                            size: :medium,
-                                            classes: "op-principal_flex"))
-        end
-        render_participant_state(participant, flex)
-      end
-    end
-
-    def render_participant_state(participant, flex)
-      if participant.attended?
-        flex.with_column(ml: 1) do
-          render(Primer::Beta::Text.new(font_size: :small, color: :subtle)) { t("description_attended").capitalize }
-        end
-      elsif participant.invited?
-        flex.with_column(ml: 1) do
-          render(Primer::Beta::Text.new(font_size: :small, color: :subtle)) { t("description_invite").capitalize }
-        end
-      end
     end
   end
 end
