+++ modules/meeting/app/components/meetings/side_panel/participants_component.html.erb
@@ -1,0 +1,56 @@
+<%=
+  component_wrapper do
+    render(Primer::OpenProject::SidePanel::Section.new) do |section|
+      section.with_title { Meeting.human_attribute_name(:participants) }
+      section.with_description { I18n.t('meeting.attachments.text') }
+      section.with_counter { @meeting.invited_or_attended_participants.count }
+
+      if @meeting.editable?
+        section.with_action_icon(
+          icon: :gear,
+          scheme: :invisible,
+          tag: :a,
+          href: participants_dialog_meeting_path(@meeting),
+          data: { controller: 'async-dialog' },
+          'aria-label': t("label_meeting_manage_participants"),
+          test_selector: "manage-participants-button",
+        )
+      end
+
+      component_wrapper(data: wrapper_data_attributes) do
+        flex_layout do |participants_container|
+          participants_container.with_row do
+            flex_layout(align_items: :center, justify_content: :space_between) do |heading|
+
+            end
+          end
+
+          participants_container.with_row(mt: 2) do
+            render(OpPrimer::ExpandableListComponent.new(cutoff_limit: 5)) do |list|
+              elements.each do |item|
+                list.with_element { render_participant(item) }
+              end
+            end
+          end
+
+          if @meeting.editable?
+            participants_container.with_row(mt: 3) do
+              render(Primer::Beta::Button.new(
+                scheme: :link,
+                color: :default,
+                underline: false,
+                font_weight: :bold,
+                tag: :a,
+                href: participants_dialog_meeting_path(@meeting),
+                data: { controller: 'async-dialog' }
+              )) do |button|
+                button.with_leading_visual_icon(icon: "person-add")
+                t("label_meeting_add_participants")
+              end
+            end
+          end
+        end
+      end
+    end
+  end
+%>
