+++ modules/meeting/app/components/meetings/side_panel/participants_dialog.html.erb
