+++ modules/meeting/app/components/meetings/side_panel/state_component.sass
