+++ modules/meeting/app/controllers/concerns/meetings/agenda_component_streams.rb
@@ -43,7 +43,7 @@ module Meetings
 
       def update_sidebar_component_via_turbo_stream(meeting: @meeting)
         update_via_turbo_stream(
-          component: Meetings::SidebarComponent.new(
+          component: Meetings::SidePanelComponent.new(
             meeting:
           )
         )
@@ -51,7 +51,7 @@ module Meetings
 
       def update_sidebar_details_component_via_turbo_stream(meeting: @meeting)
         update_via_turbo_stream(
-          component: Meetings::Sidebar::DetailsComponent.new(
+          component: Meetings::SidePanel::DetailsComponent.new(
             meeting:
           )
         )
@@ -59,7 +59,7 @@ module Meetings
 
       def update_sidebar_state_component_via_turbo_stream(meeting: @meeting)
         update_via_turbo_stream(
-          component: Meetings::Sidebar::StateComponent.new(
+          component: Meetings::SidePanel::StateComponent.new(
             meeting:
           )
         )
@@ -67,7 +67,7 @@ module Meetings
 
       def update_sidebar_details_form_component_via_turbo_stream(meeting: @meeting, status: :bad_request)
         update_via_turbo_stream(
-          component: Meetings::Sidebar::DetailsFormComponent.new(
+          component: Meetings::SidePanel::DetailsFormComponent.new(
             meeting:
           ),
           status:
@@ -76,7 +76,7 @@ module Meetings
 
       def update_sidebar_participants_component_via_turbo_stream(meeting: @meeting)
         update_via_turbo_stream(
-          component: Meetings::Sidebar::ParticipantsComponent.new(
+          component: Meetings::SidePanel::ParticipantsComponent.new(
             meeting:
           )
         )
@@ -84,7 +84,7 @@ module Meetings
 
       def update_sidebar_participants_form_component_via_turbo_stream(meeting: @meeting)
         update_via_turbo_stream(
-          component: Meetings::Sidebar::ParticipantsFormComponent.new(
+          component: Meetings::SidePanel::ParticipantsFormComponent.new(
             meeting:
           ),
           status: :bad_request # TODO: why bad_request?
