+++ modules/meeting/app/controllers/meetings_controller.rb
@@ -162,6 +162,8 @@ class MeetingsController < ApplicationController
     end
   end
 
+  def details_dialog; end
+
   def participants_dialog; end
 
   def update_participants
