+++ modules/meeting/app/views/meetings/details_dialog.turbo_stream.erb
@@ -1,0 +1,5 @@
+<%=
+  turbo_stream.dialog do
+    render(Meetings::SidePanel::DetailsDialogComponent.new(meeting: @meeting))
+  end
+%>
