+++ modules/meeting/app/views/meetings/participants_dialog.turbo_stream.erb
@@ -1,5 +1,5 @@
 <%=
   turbo_stream.dialog do
-    render(Meetings::Sidebar::ParticipantsDialog.new(meeting: @meeting))
+    render(Meetings::SidePanel::ParticipantsDialog.new(meeting: @meeting))
   end
 %>
