+++ modules/meeting/lib/open_project/meeting/engine.rb
@@ -54,7 +54,7 @@ module OpenProject::Meeting
                    contract_actions: { meetings: %i[create] }
         permission :edit_meetings,
                    {
-                     meetings: %i[edit cancel_edit update update_title update_details update_participants],
+                     meetings: %i[edit cancel_edit update update_title details_dialog update_details update_participants],
                      work_package_meetings_tab: %i[add_work_package_to_meeting_dialog add_work_package_to_meeting]
                    },
                    permissible_on: :project,
